[FIX] TeamSp API

+++ src/main/java/SSOP/ssop/controller/TeamSp/TeamSpController.java
@@ -31,6 +31,10 @@ public class TeamSpController {
     @PostMapping("/create")
     public ResponseEntity<Map<String, String>> saveTeamSp(@RequestBody TeamSp teamSp, @Login Long userId) {
         try {
+            if (userId == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body(Map.of("message", "유효한 토큰이 없습니다"));
+            }
             teamSpService.saveTeamSp(teamSp, userId); // 호스트 ID와 함께 저장
             return ResponseEntity.status(HttpStatus.CREATED)
                     .body(Map.of("message", "팀스페이스 생성 완료"));
@@ -43,8 +47,11 @@ public class TeamSpController {
     // 팀 스페이스 입장
     @PostMapping("/enter")
     public ResponseEntity<Map<String, String>> enterTeamSp(@RequestBody EnterTeamSpDto enterTeamSpDto, @Login Long userId) {
-
         try {
+            if (userId == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body(Map.of("message", "유효한 토큰이 없습니다"));
+            }
             teamSpService.EnterTeamSp(enterTeamSpDto.getInviteCode(), userId);
             return ResponseEntity.ok(Map.of("message", "팀스페이스에 성공적으로 입장하였습니다."));
         } catch (IllegalArgumentException e) {
@@ -65,7 +72,7 @@ public class TeamSpController {
 
     // 특정 팀스페이스 조회
     @GetMapping
-    public ResponseEntity<?> getTeamById(@RequestParam("team_id") Long teamId) {
+    public ResponseEntity<?> getTeamById(@RequestParam("teamId") Long teamId) {
         TeamSp teamSp = teamSpService.getTeamById(teamId);
 
         if (teamSp != null) {
@@ -78,8 +85,12 @@ public class TeamSpController {
 
     // 팀스페이스 이름 수정 (호스트만)
     @PatchMapping
-    public ResponseEntity<Map<String, String>> updateTeamSp(@RequestParam("team_id") Long teamId, @RequestBody TeamSp teamSp, @Login Long userId) {
+    public ResponseEntity<Map<String, String>> updateTeamSp(@RequestParam("teamId") Long teamId, @RequestBody TeamSp teamSp, @Login Long userId) {
         try {
+            if (userId == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body(Map.of("message", "유효한 토큰이 없습니다"));
+            }
             TeamSp updatedTeamSp = teamSpService.updateTeamSp(teamId, teamSp, userId);
             return ResponseEntity.ok(Map.of("message", "팀스페이스 이름 업데이트 완료"));
         } catch (IllegalArgumentException e) {
@@ -98,8 +109,12 @@ public class TeamSpController {
 
     // 팀스페이스 삭제 (호스트-삭제 / 참여자-퇴장)
     @DeleteMapping
-    public ResponseEntity<Map<String, String>> deleteTeamSp(@RequestParam("team_id") Long teamId, @Login Long userId) {
+    public ResponseEntity<Map<String, String>> deleteTeamSp(@RequestParam("teamId") Long teamId, @Login Long userId) {
         try {
+            if (userId == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body(Map.of("message", "유효한 토큰이 없습니다"));
+            }
             teamSpService.deleteTeamSp(teamId, userId);
             return ResponseEntity.ok(Map.of("message", "팀스페이스가 삭제되었습니다."));
         } catch (IllegalArgumentException e) {
